fix: Fix linter tests

+++ solution/include/image.h
@@ -12,7 +12,7 @@ struct image {
 };
 
 struct image image_create(const struct dimensions size);
-size_t image_get_padding_in_bytes(const struct image *image);
+long image_get_padding_in_bytes(const struct image *image);
 void image_destroy(struct image *image);
 size_t image_get_width(const struct image *img);
 size_t image_get_height(const struct image *img);
